doto-graph/requirements.txt
caldav>=0.6.2
flask>=1.1.1
Flask-GraphQL>=2.0.1
flask-cors>=3.0.8
google-api-python-client>=1.7.11
google-auth-oauthlib>=0.4.1
graphene>=2.1.8
waitress>=1.4.3
